Refactor: split Ytdlp.Fluent.cs into partial classes
Refactored the codebase to split the large Ytdlp.Fluent.cs file into multiple logically grouped partial class files (e.g., General, Authentication, Download, etc.), improving maintainability and IntelliSense support. Updated README.md to reflect the new structure and added more options/overloads. Introduced YtdlpPreset enum for strongly-typed preset aliases. Moved AddFlag and AddOption to the main Ytdlp class for consistent option handling. Reformatted event region and improved XML docs. No functional changes; this is an organizational refactor.
